Plot the networks, using different colors for transcription factors, genes of interest, and additional genes.

Plot the networks, using different colors for transcription factors, genes of interest, and additional genes.

Usage

PlotNetwork(
  network,
  genesOfInterest,
  tfColor = "blue",
  nodeSize = 1,
  edgeWidth = 0.5,
  vertexLabels = NA,
  vertexLabelSize = 0.7,
  vertexLabelOffset = 0.5,
  layoutBipartite = TRUE,
  geneColorMapping = NULL
)

Arguments

network

A data frame with the following format: tf,gene

genesOfInterest

Which genes of interest to highlight

tfColor

Color for the transcription factors

nodeSize

Size of node

edgeWidth

Width of edges

vertexLabels

Which vertex labels to include. By default, none are included.

vertexLabelSize

The size of label to use for the vertex, as a fraction of the default.

vertexLabelOffset

Number of pixels in the offset when plotting labels. Default is TRUE.

layoutBipartite

Whether or not to layout as a bipartite graph.

geneColorMapping

Color mapping from a set of genes to a color. The nodes and edges connected to them will be this color. If NULL, all genes and their edges will be gray. The format is a data frame, where the first column ("gene") is the name of the gene and the second ("color") is the color.

Value

A bipartite plot of the network