Edge subtraction method (CONDOR optimizaton)

Takes two networks, subtracts edges individually, and then clusters the subtracted network using CONDOR.

alpaca.DeltaZAnalysis(net.table, file.stem)

Arguments

net.table
A table of edges, with the first column representing the TFs ("from" nodes) and the second column representing the targets ("to" nodes). The third column contains the edge weights corresponding to the control or healthy network, and the fourth column contains the edge weights for the disease network or network of interest.

file.stem
The folder location and title under which all results will be stored.

Value

List where first element is the membership vector and second element is the contribution score of each node to its community's modularity in the final edge-subtracted network