Calculate Qscore for all nodes

Qscore is designed to calculate the fraction of the modularity contributed by each node to its community's modularity

condor.qscore(condor.object)

Arguments

condor.object
output of condor.cluster or condor.modularity.max

Value

condor.object list has condor.object$qscores added to it. this includes two data.frames, blue.qscore and red.qscore which have the qscore for each red and blue node.

Examples

r = c(1,1,1,2,2,2,3,3,3,4,4);
b = c(1,2,3,1,2,4,2,3,4,3,4);
reds <- c("Alice","Sue","Janine","Mary")
blues <- c("Bob","John","Ed","Hank")
elist <- data.frame(red=reds[r],blue=blues[b])
condor.object <- create.condor.object(elist)
condor.object <- condor.cluster(condor.object)

#> [1] "modularity of projected graph 0"
#> [1] "Q = 0"
#> [1] "Q = 0.198347107438017"
#> [1] "Q = 0.231404958677686"
#> [1] "Q = 0.231404958677686"

condor.object <- condor.qscore(condor.object)