Bipartite Edge Reconstruction from Expression data

This function generates a complete bipartite network from gene expression data and sequence motif data

monster.monsterNI(
  motif.data,
  expr.data,
  verbose = FALSE,
  randomize = "none",
  method = "bere",
  ni.coefficient.cutoff = NA,
  alphaw = 1,
  regularization = "none",
  score = "motifincluded",
  cpp = FALSE
)

Arguments

expr.data
An expression dataset, as a genes (rows) by samples (columns)

verbose
logical to indicate printing of output for algorithm progress.

randomize
logical indicating randomization by genes, within genes or none

method
String to indicate algorithm method. Must be one of "bere","pearson","cd","lda", or "wcd". Default is "bere". Important note: the direct regulatory network observed from gene expression is currently implemented as a regular correlation as opposed to the partial correlation described in the paper (please see Schlauch et al., 2017, https://doi.org/10.1186/s12918-017-0517-y)

ni.coefficient.cutoff
numeric to specify a p-value cutoff at the network inference step. Default is NA, indicating inclusion of all coefficients.

alphaw
A weight parameter specifying proportion of weight to give to indirect compared to direct evidence. See documentation.

regularization
String parameter indicating one of "none", "L1", "L2"

score
String to indicate whether motif information will be readded upon completion of the algorithm

cpp
logical use C++ for maximum speed, set to false if unable to run.

motif
A motif dataset, a data.frame, matrix or exprSet containing 3 columns. Each row describes an motif associated with a transcription factor (column 1) a gene (column 2) and a score (column 3) for the motif.

Value

matrix for inferred network between TFs and genes

Examples

data(yeast)
cc.net <- monster.monsterNI(yeast$motif,yeast$exp.cc)

#> Time difference of 1.14611 secs
#> Time difference of 0.4518781 secs