Bi-partite network analysis tools

This function analyzes a bi-partite network.

monster.transformation.matrix(network.1, network.2, by.tfs = TRUE,
  standardize = FALSE, remove.diagonal = TRUE, method = "ols")

Arguments

network.1
starting network, a genes by transcription factors data.frame with scores for the existence of edges between

network.2
final network, a genes by transcription factors data.frame with scores for the existence of edges between

by.tfs
logical indicating a transcription factor based transformation. If false, gives gene by gene transformation matrix

standardize
logical indicating whether to standardize the rows and columns

remove.diagonal
logical for returning a result containing 0s across the diagonal

method
character specifying which algorithm to use, default='ols'

Value

matrix object corresponding to transition matrix

Examples

# \donttest{
data(yeast)
cc.net.1 <- monster.monsterNI(yeast$motif,yeast$exp.cc[1:1000,1:20])

#> Time difference of 0.075068 secs
#> .

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> ......................

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> ..........

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> ....................

cc.net.2 <- monster.monsterNI(yeast$motif,yeast$exp.cc[1:1000,31:50])

#> Time difference of 0.07555008 secs
#> .......................

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> ..............................

monster.transformation.matrix(cc.net.1, cc.net.2)

#> Error in svd(X): infinite or missing values in 'x'

# }