Bipartite Edge Reconstruction from Expression data (composite method with direct/indirect)

This function generates a complete bipartite network from gene expression data and sequence motif data. This NI method serves as a default method for inferring bipartite networks in MONSTER. Running monsterBereFull can generate these networks independently from the larger MONSTER method.

monsterBereFull(
  motif.data,
  expr.data,
  alpha = 0.5,
  lambda = 10,
  score = "motifincluded"
)

Arguments

motif.data
A motif dataset, a data.frame, matrix or exprSet containing 3 columns. Each row describes an motif associated with a transcription factor (column 1) a gene (column 2) and a score (column 3) for the motif.

expr.data
An expression dataset, as a genes (rows) by samples (columns) data.frame

alpha
A weight parameter specifying proportion of weight to give to indirect compared to direct evidence. See documentation.

lambda
if using penalized, the lambda parameter in the penalized logistic regression

score
String to indicate whether motif information will be readded upon completion of the algorithm

Value

An matrix or data.frame

Examples

data(yeast)
monsterRes <- monsterBereFull(yeast$motif, yeast$exp.cc, alpha=.5)

#> Using V3 as value column: use value.var to override.

#> .123456
#> .12345
#> .12345
#> .12345
#> .12345
#> .123456
#> .12345
#> .12345
#> .12345
#> .123456
#> .12345
#> .12345
#> .12345
#> .12345
#> .12345
#> .123456
#> .1234
#> .12345
#> .12345
#> .12345
#> .1234
#> .12345
#> .123456
#> .12345
#> .12345
#> .1234
#> .1234
#> .12345
#> .12345
#> .12345
#> .12345
#> .12345
#> .123456
#> .12345
#> .12345
#> .123456
#> .12345
#> .12345
#> .12345
#> .12345
#> .12345
#> .12345
#> .1234
#> .12345
#> .12345
#> .12345
#> .12345
#> .12345
#> .12345
#> .12345
#> .1234
#> .12345
#> .12345