Run OTTER in R

Description: OTTER infers gene regulatory networks using TF DNA binding motif (W), TF PPI (P), and gene coexpression (C) through minimzing the following objective: min f(W) with f(W) = (1-lambda)*||WW' - P||^2 + lambda*||W'W - C||^2 + (gamma/2)*||W||^2

otter(
  W,
  P,
  C,
  lambda = 0.0035,
  gamma = 0.335,
  Iter = 32,
  eta = 1e-05,
  bexp = 1
)

Arguments

W
: TF-gene regulatory network based on TF motifs as a matrix of size (t,g), g=number of genes, t=number of TFs

P
: TF-TF protein interaction network as a matrix of size (t,t)

C
: gene coexpression as a matrix of size (g,g)

lambda
: tuning parameter in [0,1] (higher gives more weight to C)

gamma
: regularization parameter

Iter
: number of iterations of the algorithm

eta
: learning rate

bexp
: exponent influencing learning rate (higher means smaller)

Outputs:

Value

W : Predicted TF-gene complete regulatory network as an adjacency matrix of size (t,g).

Details

Inputs:

Examples

W=matrix(rexp(100, rate=.1), ncol=10)
C=matrix(rexp(100, rate=.1), ncol=10)
P=matrix(rexp(100, rate=.1), ncol=10)

# Run OTTER algorithm
W <- otter(W, P, C)