TIGER main function

TIGER main function

Usage

tiger(
  expr,
  prior,
  method = "VB",
  TFexpressed = TRUE,
  signed = TRUE,
  baseline = TRUE,
  psis_loo = FALSE,
  seed = 123,
  out_path = NULL,
  out_size = 300,
  a_sigma = 1,
  b_sigma = 1,
  a_alpha = 1,
  b_alpha = 1,
  sigmaZ = 10,
  sigmaB = 1,
  tol = 0.005
)

Arguments

expr

A normalized log-transformed gene expressison matrix. Rows are genes and columns are sampeles (cells).

prior

A prior regulatory network in adjacency matrix format. Rows are TFs and columns target genes.

method

Method used for Bayesian inference. "VB" or "MCMC". Defaults to "VB".

TFexpressed

TF mRNA needs to be expressed or not. Defaults to TRUE.

signed

Prior network is signed or not. Defaults to TRUE.

baseline

Include baseline or not. Defaults to TRUE.

psis_loo

Use pareto smoothed importance sampling leave-one-out cross validation to check model fitting or not. Defaults to FALSE.

seed

Seed for reproducible results. Defaults to 123.

out_path

(Optional) output path for CmdStanVB or CmdStanMCMC object. Defaults to NULL.

out_size

Posterior sampling size. Default = 300.

a_sigma

Hyperparameter of error term. Default = 1.

b_sigma

Hyperparameter of error term. Default = 1.

a_alpha

Hyperparameter of edge weight W. Default = 1.

b_alpha

Hyperparameter of edge weight W. Default = 1.

sigmaZ

Standard deviation of TF activity Z. Default = 10.

sigmaB

Standard deviation of baseline term. Default = 1.

tol

Convergence tolerance on ELBO.. Default = 0.005.

Value

A TIGER list object. * W is the estimated regulatory network, but different from prior network, rows are genes and columns are TFs. * Z is the estimated TF activities, rows are TFs and columns are samples. * TF.name, TG.name, and sample.name are the used TFs, target genes and samples. * If psis_loo is TRUE, loocv is a table of psis_loo result for model checking. * If psis_loo is TRUE, elpd_loo is the Bayesian LOO estimate of the expected log pointwise predictive density, which can be used for Bayesian stacking to handle multi-modality later.

Examples

data(TIGER_expr)
data(TIGER_prior)
tiger(TIGER_expr,TIGER_prior)