Bi-partite network analysis tools

This function analyzes a bi-partite network.

transformation.matrix(network.1, network.2, by.tfs = TRUE,
  standardize = FALSE, remove.diagonal = TRUE, method = "ols")

Arguments

network.1
starting network, a genes by transcription factors data.frame with scores for the existence of edges between

network.2
final network, a genes by transcription factors data.frame with scores for the existence of edges between

by.tfs
logical indicating a transcription factor based transformation. If false, gives gene by gene transformation matrix

standardize
logical indicating whether to standardize the rows and columns

remove.diagonal
logical for returning a result containing 0s across the diagonal

method
character specifying which algorithm to use, default='ols'

Value

matrix object corresponding to transition matrix

Examples

data(yeast)
cc.net.1 <- monsterNI(yeast$motif,yeast$exp.cc[1:1000,1:20])

#> Time difference of 0.0792141 secs
#> Time difference of 0.106308 secs

cc.net.2 <- monsterNI(yeast$motif,yeast$exp.cc[1:1000,31:50])

#> Time difference of 0.1400521 secs
#> Time difference of 0.08625603 secs

transformation.matrix(cc.net.1, cc.net.2)

#> [1] "Using OLS method"

#>               YBL005W       YBR083W       YBR240C       YCR106W       YDL056W
#> YBL005W  0.0000000000 -5.230999e-02 -4.616050e-02 -0.0568845578  3.434095e-02
#> YBR083W  0.0144183765  0.000000e+00  1.844380e-04 -0.0103175357  4.212441e-03
#> YBR240C  0.0286556899 -4.041948e-03  0.000000e+00  0.0137787070  2.231064e-02
#> YCR106W -0.0201408907 -7.723884e-03  5.970660e-03  0.0000000000  1.914782e-02
#> YDL056W -0.0560805414  4.951251e-03  6.020824e-03  0.0021343777  0.000000e+00
#> YDL106C  0.0424669279  3.742008e-02  1.018351e-02  0.0084968622 -2.166975e-02
#> YDR216W -0.0018699959 -9.340690e-03  1.002211e-02 -0.0117849280 -4.246789e-02
#> YDR259C -0.0406635934 -1.716916e-03 -2.147080e-03  0.0100434770 -1.244487e-02
#> YDR423C  0.0767905585 -2.135990e-03  5.939197e-03 -0.0193894677 -2.083356e-03
#> YDR463W  0.1393352022  9.064398e-03  9.036819e-02 -0.0819559983 -2.061533e-01
#> YEL009C -0.0226736076 -5.447868e-03  3.040100e-03 -0.0025789650 -1.767212e-02
#> YER040W -0.0182769147 -6.357523e-03  2.556057e-03  0.0027916695 -8.026079e-03
#> YFR034C -0.0540984234 -8.347687e-04 -9.512529e-03  0.0139342004  2.364606e-02
#> YGL013C  0.0637294578  6.220969e-03  6.899688e-03  0.0054195582  5.402310e-03
#> YGL162W  0.0146425924 -4.833541e-03  2.304158e-02 -0.0183259296 -3.292693e-02
#> YHL020C  0.1890860931  1.430196e-02  3.707995e-05 -0.0441352482 -4.489540e-02
#> YHR084W  0.0195208968  2.561459e-03  1.303267e-03  0.0029949402 -1.067964e-03
#> YHR178W -0.0262461880  3.019337e-03 -1.707177e-02  0.0008835262  6.510163e-02
#> YHR206W -0.0397060042  3.336150e-03 -7.773902e-03  0.0052480324 -7.364103e-04
#> YIL101C -0.0535338298  1.077476e-02  1.730384e-02  0.0135935930 -3.440038e-02
#> YIL131C -0.0322681656 -2.253221e-03 -1.022357e-02 -0.0002870155 -4.247828e-03
#> YIR018W -0.0406635934 -1.716916e-03 -2.147080e-03  0.0100434770 -1.244487e-02
#> YJL056C  0.1382818397 -2.106116e-02  1.667470e-02 -0.1099145983  1.919490e-01
#> YJL110C  0.0044341406  5.520965e-03 -6.065923e-03  0.0006702261  3.260280e-02
#> YJR060W  0.1133975759 -4.442973e-04  2.065291e-02 -0.0311638481 -2.264913e-02
#> YKL038W  0.0241903275  1.343298e-03  4.895518e-03  0.0070616779 -8.758221e-03
#> YKL043W  0.0202344506  8.130379e-04 -2.271992e-03  0.0052070187  8.890186e-03
#> YKL062W  0.0700566039 -4.330762e-03  2.701241e-02 -0.0451691494 -2.291198e-02
#> YKL109W  0.0304954817 -9.175742e-05  1.143720e-02 -0.0016432970 -2.723507e-02
#> YKR099W  0.0102696396 -1.317026e-03  8.598416e-03 -0.0183907603 -1.734007e-02
#> YLR131C  0.0358549942 -2.434898e-03 -8.260895e-03 -0.0017277406  3.645206e-02
#> YLR182W  0.0268018776  6.446407e-03  1.709050e-03 -0.0020623433 -1.331319e-02
#> YLR228C  0.0506934501 -6.615131e-02 -5.515320e-02  0.0698591206 -5.102204e-02
#> YLR403W -0.0061048296  1.192649e-03 -8.021694e-03  0.0012639202  2.047572e-02
#> YLR451W -0.0790758092 -1.151093e-02  3.544036e-02 -0.0670713542  9.962801e-02
#> YML051W  0.0273105352  2.276428e-03 -3.401896e-02  0.0362530026  7.748882e-02
#> YMR016C -0.0409114436 -9.652666e-03 -7.841105e-03  0.0106583667 -1.604565e-03
#> YMR019W -0.0779824376 -8.593776e-03  4.911207e-04  0.0122971904  9.482333e-03
#> YMR043W  0.0713110843 -4.587445e-03  2.052738e-02 -0.0219627068 -3.760898e-02
#> YMR070W -0.0045939062 -8.026793e-04 -7.497028e-04  0.0037067703  7.658473e-05
#> YNL068C -0.0597315652 -7.293381e-04 -9.316022e-03  0.0224883371 -1.732009e-02
#> YNL216W  0.0038965321  9.209643e-03 -1.896673e-02  0.0124799225  5.988019e-02
#> YOL028C  0.0055742288 -2.384688e-03 -5.450544e-03  0.0048893281  5.448411e-03
#> YOR113W -0.0521701597  2.486574e-03  3.330797e-03  0.0079168379 -2.620034e-02
#> YOR344C -0.0133216742  1.438303e-02 -1.184609e-02  0.0126530996  2.067166e-02
#> YPL038W  0.0530569647 -4.381904e-03  2.096492e-02 -0.0310695246 -1.505342e-02
#> YPL049C -0.0352346895 -1.125171e-02 -1.940801e-02  0.0090314534  1.694385e-02
#> YPL075W  0.0367594475  1.507981e-02 -7.001718e-03 -0.0088460740 -7.187532e-05
#> YPL089C -0.1167066269 -1.050356e-02 -5.501799e-02  0.0155848698  1.604676e-01
#> YPL139C  0.0007695277  2.842564e-03  6.031204e-03  0.0034237087 -1.669022e-02
#> YPL202C  0.0175605577 -2.457854e-04 -4.620149e-03  0.0061763142 -5.504239e-03
#> YPR065W  0.0419385642  9.783317e-03  6.067574e-03 -0.0071485622 -1.260450e-02
#> YPR104C  0.0795906366 -7.425639e-03  2.494978e-02 -0.0281857599 -6.306044e-02
#>               YDL106C       YDR216W       YDR259C       YDR423C       YDR463W
#> YBL005W -0.0433338841 -2.714221e-02 -1.847008e-03 -0.1064587253 -1.614088e-01
#> YBR083W  0.0092270807 -1.187019e-02 -1.008756e-02 -0.0074981626 -9.327951e-03
#> YBR240C  0.0316134883  9.603036e-03 -3.144999e-03 -0.0032623662  3.889565e-02
#> YCR106W  0.0043948758 -3.296298e-02  5.575047e-03 -0.0112178787 -2.961687e-02
#> YDL056W -0.0256929930 -5.143635e-03 -1.860046e-02  0.0028003531 -5.976815e-02
#> YDL106C  0.0000000000  2.229839e-02  1.203688e-02  0.0200209990  2.254115e-02
#> YDR216W -0.0095899425  0.000000e+00 -1.781547e-03  0.0086908539 -1.058347e-02
#> YDR259C -0.0147709469  8.474724e-03  0.000000e+00  0.0026755101 -7.011194e-03
#> YDR423C  0.0303463652 -5.706661e-03 -9.825120e-03  0.0000000000 -1.025530e-02
#> YDR463W  0.0605572194 -5.426080e-02 -5.377891e-02  0.0164852446  0.000000e+00
#> YEL009C -0.0117666207  2.021544e-03 -6.109212e-03  0.0063942172 -9.757279e-03
#> YER040W  0.0094926230  2.879602e-03 -4.185205e-04 -0.0078377409  1.525723e-02
#> YFR034C -0.0082047130 -3.600504e-03  9.643882e-03 -0.0162371417  5.708774e-03
#> YGL013C  0.0148587062  3.382059e-03  4.068630e-03  0.0079441508  3.051694e-02
#> YGL162W  0.0096186107 -3.576236e-03 -9.385802e-03  0.0121932414  1.589653e-02
#> YHL020C  0.0466315259  6.454111e-03 -3.683352e-02  0.0219203477 -2.554400e-02
#> YHR084W  0.0026027915  2.211713e-06  6.709671e-03  0.0053568772  9.988817e-03
#> YHR178W  0.0150962826 -4.184645e-03  2.674635e-04 -0.0172048594  3.041784e-02
#> YHR206W -0.0368849595  2.941204e-03 -4.452890e-04  0.0019036431 -8.453165e-03
#> YIL101C -0.0707452435  1.926750e-02 -1.809722e-02  0.0198628271  1.491815e-03
#> YIL131C -0.0178500126 -6.090762e-03 -1.073849e-02 -0.0112892450 -3.892264e-05
#> YIR018W -0.0147709469  8.474724e-03  2.001376e-02  0.0026755101 -7.011194e-03
#> YJL056C -0.0760714689  2.654399e-02 -1.841293e-02  0.0284560650 -4.501102e-02
#> YJL110C -0.0004081917 -1.113362e-02  4.804594e-03 -0.0049862042 -2.250832e-02
#> YJR060W  0.0591970291 -1.802036e-02 -1.999667e-02  0.0101520076  8.272268e-03
#> YKL038W  0.0219864508  5.787699e-03 -9.590663e-03  0.0006405852  1.762993e-02
#> YKL043W  0.0091359122  2.516309e-03 -8.766494e-04  0.0011232202  2.495975e-03
#> YKL062W  0.0259184805 -3.949852e-02  3.366246e-05  0.0029205134 -6.520269e-02
#> YKL109W  0.0146038827  1.343365e-03 -9.571792e-04  0.0119881378 -7.570029e-04
#> YKR099W -0.0008946890 -1.118061e-02 -6.642811e-03  0.0074130100 -2.333466e-02
#> YLR131C  0.0386876802 -1.909706e-02  2.247435e-02 -0.0165913669 -3.355873e-03
#> YLR182W -0.0066634021 -5.470277e-06 -1.133557e-02  0.0046130042 -9.891891e-03
#> YLR228C  0.0699931195  2.875974e-04 -5.912152e-02  0.0309006737  1.065945e-02
#> YLR403W -0.0126441192 -6.768187e-03  5.142883e-03 -0.0029359333 -1.659923e-03
#> YLR451W -0.0195645633 -8.304461e-02  6.970872e-02 -0.0588704968 -8.498039e-02
#> YML051W -0.0038485969  2.211181e-02  1.828614e-02 -0.0319575032  4.657348e-02
#> YMR016C  0.0033255650 -4.064585e-04  9.755221e-03 -0.0092863295 -1.615186e-03
#> YMR019W -0.0774138213 -9.434251e-03  2.401556e-02 -0.0115335531  8.005984e-02
#> YMR043W  0.0445507565 -4.934745e-03 -3.680956e-03  0.0198939718 -2.894582e-02
#> YMR070W  0.0068735722 -3.667471e-03  6.723144e-03  0.0023749144 -5.344173e-03
#> YNL068C -0.0368565410  1.730706e-02  1.157747e-02  0.0023903845  1.665001e-02
#> YNL216W  0.0124362920  1.034219e-02 -1.722546e-03 -0.0206724953  5.511245e-02
#> YOL028C  0.0061249302 -3.112863e-03  7.994221e-03 -0.0055760944  2.401476e-02
#> YOR113W -0.0309699434  2.594759e-02 -8.245089e-03  0.0052756414  2.394193e-02
#> YOR344C -0.0265792227  1.236378e-02 -1.437563e-04 -0.0080297353  2.948360e-03
#> YPL038W  0.0536375841 -2.582677e-02 -1.485165e-02  0.0020782888 -2.407991e-02
#> YPL049C -0.0100452779 -1.318010e-02  1.561530e-03 -0.0140233251 -8.400824e-03
#> YPL075W -0.0181401438  5.412480e-03 -2.522902e-03  0.0049713596 -1.411619e-02
#> YPL089C  0.0018509798 -1.720198e-02  5.755546e-02 -0.0592420197  1.573962e-02
#> YPL139C -0.0104766036  1.098220e-05  3.389149e-03  0.0035489159  9.123440e-03
#> YPL202C  0.0252388223  8.019504e-03 -4.670069e-04  0.0001920762  1.373445e-02
#> YPR065W -0.0148394108  8.114509e-03 -4.337653e-03  0.0079395088 -8.246101e-03
#> YPR104C -0.0111897743 -1.574921e-02 -1.295211e-02  0.0291412721 -5.369333e-02
#>               YEL009C       YER040W       YFR034C       YGL013C       YGL162W
#> YBL005W  0.0326344382 -8.508187e-04 -0.0030614990 -1.905714e-02 -0.1039046767
#> YBR083W -0.0047683284 -2.211924e-03  0.0005564389  2.880117e-04 -0.0102576900
#> YBR240C -0.0008855613 -5.025724e-03  0.0060834545  1.832018e-02  0.0231532174
#> YCR106W -0.0132611319 -4.590003e-03 -0.0065911953 -9.994493e-03 -0.0341919984
#> YDL056W  0.0130607265  2.103779e-02  0.0004853351 -1.363234e-02 -0.0109042729
#> YDL106C  0.0077790619  2.537362e-02  0.0169477115 -7.433333e-05  0.0293345584
#> YDR216W  0.0051031714  1.114190e-03 -0.0142780860 -1.196622e-02 -0.0113538180
#> YDR259C  0.0081059878  7.143453e-03  0.0063430255 -1.119420e-02  0.0046220271
#> YDR423C -0.0051717586 -1.947049e-02 -0.0159773987  1.352036e-02  0.0033090781
#> YDR463W  0.0016290163  3.499920e-03 -0.0368480042  3.068215e-03 -0.0422608935
#> YEL009C  0.0000000000  1.087473e-03 -0.0016464918 -7.358135e-03 -0.0009320577
#> YER040W  0.0029439789  0.000000e+00  0.0050431251  3.576370e-03  0.0062063098
#> YFR034C -0.0046260434  6.268064e-03  0.0000000000 -5.171076e-03 -0.0012821946
#> YGL013C -0.0034674962 -3.928294e-04  0.0027938727  0.000000e+00  0.0171244143
#> YGL162W  0.0078399909  6.659993e-03 -0.0090188476 -4.078032e-03  0.0000000000
#> YHL020C  0.0073233621 -6.550786e-03 -0.0146552729  1.366908e-02  0.0050390101
#> YHR084W -0.0042047464  1.803398e-03  0.0037889680  2.247028e-03  0.0022470788
#> YHR178W -0.0107397587 -1.085453e-02  0.0039484578 -9.287208e-03  0.0163095996
#> YHR206W  0.0028698323  4.942784e-03 -0.0001713253  7.018884e-04 -0.0001041712
#> YIL101C  0.0255280118  2.293802e-02  0.0013883943 -1.888170e-02  0.0113944963
#> YIL131C -0.0001038235  2.436667e-03  0.0024448596 -5.055260e-03 -0.0060163140
#> YIR018W  0.0081059878  7.143453e-03  0.0063430255 -1.119420e-02  0.0046220271
#> YJL056C  0.0028399598 -1.391340e-02  0.0126214739 -1.404329e-02 -0.0015428171
#> YJL110C -0.0110031836 -4.380547e-03 -0.0016990596 -6.398624e-03 -0.0096194037
#> YJR060W -0.0089381734 -9.685795e-03 -0.0126392392  1.850673e-02 -0.0166752535
#> YKL038W  0.0067837341  3.620529e-03  0.0048578689  1.298405e-02  0.0048019403
#> YKL043W  0.0002759888 -1.927975e-03  0.0028329706  7.351244e-03 -0.0011659848
#> YKL062W -0.0174440265 -1.438754e-02 -0.0256846691 -6.575487e-04 -0.0377755444
#> YKL109W  0.0014199279 -2.193287e-03 -0.0072729119  6.833863e-03  0.0015976876
#> YKR099W -0.0002654401  2.987898e-03 -0.0082048352 -1.228554e-03 -0.0155106830
#> YLR131C -0.0223326314 -1.070358e-02 -0.0065771087  4.520432e-03 -0.0186215755
#> YLR182W  0.0046342206  5.751119e-03  0.0049003374  7.409289e-03 -0.0051512295
#> YLR228C -0.0279538442 -4.188579e-03 -0.0037588090  3.308412e-02 -0.0455763782
#> YLR403W -0.0074687734 -1.497762e-05 -0.0038498520 -2.691749e-03 -0.0106900969
#> YLR451W -0.0675462350 -5.536092e-02 -0.0499660249 -4.852511e-02 -0.0848093577
#> YML051W -0.0112062852 -1.334809e-02  0.0207661737  2.991044e-02  0.0335682133
#> YMR016C  0.0032152479 -3.566875e-04  0.0058004575 -3.697728e-03  0.0004279883
#> YMR019W -0.0108308584  6.356761e-03 -0.0192173479 -1.439673e-02 -0.0136982612
#> YMR043W  0.0012833811 -1.682882e-02 -0.0238174754  1.376344e-03 -0.0089853195
#> YMR070W  0.0007812560  2.712867e-04 -0.0018233815  5.182669e-03 -0.0124774202
#> YNL068C  0.0105097119  1.083033e-02  0.0052967090 -9.958339e-03  0.0192646625
#> YNL216W -0.0119815907  8.544734e-04  0.0165561629  3.471744e-03  0.0159892184
#> YOL028C -0.0040591496  1.431978e-04  0.0032569946  8.470719e-03  0.0002112734
#> YOR113W  0.0179673008  9.026798e-03  0.0124812217 -8.636810e-05  0.0183508258
#> YOR344C  0.0033273255  6.755135e-03  0.0093994904 -4.036458e-03  0.0181075634
#> YPL038W -0.0063435588 -1.954543e-03 -0.0082041615  1.692292e-02 -0.0399981869
#> YPL049C -0.0086048448 -1.251916e-02 -0.0031671394 -1.708217e-03 -0.0069714817
#> YPL075W -0.0034234145 -1.589631e-04  0.0056106641  1.170541e-03  0.0050366213
#> YPL089C -0.0558110303 -4.575323e-02  0.0155401940 -3.144967e-02  0.0244290392
#> YPL139C  0.0009665034  3.266842e-03 -0.0046217079 -9.281204e-04 -0.0011122163
#> YPL202C  0.0019961730  1.623910e-03  0.0033170782  2.432855e-03  0.0092704075
#> YPR065W  0.0065146124 -1.770260e-03 -0.0002530540  8.171392e-03  0.0053944573
#> YPR104C  0.0109830552 -3.612711e-03 -0.0239880458  9.631757e-03 -0.0250324982
#>               YHL020C       YHR084W       YHR178W       YHR206W       YIL101C
#> YBL005W  0.1399446235 -0.0854077380 -0.1204923948 -0.1191798813 -0.2012858926
#> YBR083W  0.0121484499 -0.0056278235  0.0128233235 -0.0062351282  0.0118718176
#> YBR240C  0.0023206704 -0.0102260742 -0.0111792299  0.0198217567  0.0327041658
#> YCR106W -0.0291184678  0.0002199986 -0.0062274129 -0.0150092996  0.0177182681
#> YDL056W -0.0196351567  0.0077686101  0.0326778671 -0.0292689816 -0.0085475014
#> YDL106C  0.0234353234 -0.0018805283  0.0169482742 -0.0185070472 -0.0335294009
#> YDR216W  0.0025627399 -0.0062252557 -0.0010725206 -0.0079426404  0.0010435483
#> YDR259C -0.0095190476  0.0002927657 -0.0006446023 -0.0028766144 -0.0047358058
#> YDR423C  0.0303623027 -0.0078137071  0.0153122965  0.0058415281  0.0111810816
#> YDR463W  0.0022778900  0.0014552109  0.1032110414 -0.0680986509 -0.0043715600
#> YEL009C  0.0045316135 -0.0068656442  0.0049634518 -0.0045446144  0.0062686885
#> YER040W  0.0030908451 -0.0090279550 -0.0127562025  0.0109098794  0.0206048551
#> YFR034C -0.0122760434  0.0028389476 -0.0103533469 -0.0040014913  0.0005574415
#> YGL013C  0.0014947134  0.0034844383 -0.0134856991  0.0091073187 -0.0116443542
#> YGL162W  0.0128400053 -0.0017240435  0.0207715329 -0.0140787674  0.0090161404
#> YHL020C  0.0000000000  0.0082523697  0.0688937951 -0.0004643036  0.0257736865
#> YHR084W -0.0078855352  0.0000000000 -0.0085330322 -0.0024808929 -0.0105767101
#> YHR178W  0.0385192487 -0.0264110200  0.0000000000  0.0106825123  0.0268672751
#> YHR206W -0.0023954956 -0.0035242631 -0.0113657603  0.0000000000 -0.0230135921
#> YIL101C  0.0109086318  0.0003725988  0.0146140591 -0.0147043049  0.0000000000
#> YIL131C  0.0011002021 -0.0053579258 -0.0033483669 -0.0030024352 -0.0017311409
#> YIR018W -0.0095190476  0.0002927657 -0.0006446023 -0.0028766144 -0.0047358058
#> YJL056C  0.0166311676  0.0281155815  0.0325194038 -0.0495867830  0.0357005070
#> YJL110C -0.0190014542  0.0078029044  0.0117585193 -0.0121113332 -0.0128166440
#> YJR060W  0.0271453067 -0.0057246986  0.0117454933  0.0004846110  0.0113414413
#> YKL038W  0.0055547229  0.0016598433  0.0165925840  0.0075471555  0.0161386512
#> YKL043W -0.0046783963  0.0031327112  0.0069517236  0.0037212320  0.0066195912
#> YKL062W -0.0069183020 -0.0055378352  0.0180813842 -0.0241303254 -0.0064811663
#> YKL109W  0.0060253490  0.0022865447 -0.0007485313 -0.0010895294 -0.0179708978
#> YKR099W  0.0058904684 -0.0065428514  0.0101564814 -0.0160282907 -0.0004086964
#> YLR131C -0.0336807541  0.0160751348 -0.0198799282  0.0010008135 -0.0106446346
#> YLR182W  0.0006373933  0.0082544144  0.0179081129 -0.0023435596 -0.0101739469
#> YLR228C -0.1652129759  0.0900056973 -0.0992422800  0.0256416653  0.1000447246
#> YLR403W -0.0085061102  0.0038909687 -0.0052446388 -0.0068796182 -0.0121993042
#> YLR451W -0.0051127170  0.0102162209 -0.0207504157 -0.0232737488 -0.0437429775
#> YML051W  0.0131738096  0.0106560404 -0.0495500234  0.0359332121  0.0254456513
#> YMR016C -0.0211202184  0.0058329509 -0.0160261142  0.0037304857 -0.0070480629
#> YMR019W -0.0169222263 -0.0141898861 -0.0629077699  0.0140128285  0.0280689761
#> YMR043W  0.0329933543 -0.0072268004  0.0022127434 -0.0094901110 -0.0146382825
#> YMR070W -0.0104847938  0.0030449092 -0.0083687895 -0.0035263714 -0.0133045085
#> YNL068C -0.0187546461  0.0049141936 -0.0218367668  0.0095512901 -0.0045576112
#> YNL216W  0.0266537931  0.0016947619  0.0033753370  0.0195788557  0.0255923921
#> YOL028C -0.0144224584  0.0010775661 -0.0232443378  0.0083513721  0.0054526860
#> YOR113W  0.0191002435  0.0022257312 -0.0095002615  0.0119165985 -0.0025794904
#> YOR344C  0.0084161580  0.0052283641  0.0139712592  0.0043682442 -0.0090943364
#> YPL038W  0.0244705718 -0.0055127826  0.0238081166 -0.0175910659  0.0121261556
#> YPL049C  0.0148294303 -0.0101069987 -0.0261885544 -0.0045666334 -0.0156551424
#> YPL075W  0.0024114278  0.0027500954  0.0001254865 -0.0026393906 -0.0142598650
#> YPL089C  0.0051849251 -0.0203338663 -0.0850271048  0.0257669594  0.0118730056
#> YPL139C -0.0032064229  0.0052413657 -0.0023557679 -0.0020518757 -0.0099655516
#> YPL202C  0.0051439785  0.0055702980 -0.0006580483  0.0034120283  0.0003360935
#> YPR065W  0.0004466608  0.0076397377  0.0074232610  0.0055628874 -0.0126735555
#> YPR104C -0.0126983965 -0.0113260472  0.0077556458 -0.0204005348 -0.0080490499
#>               YIL131C       YIR018W       YJL056C       YJL110C       YJR060W
#> YBL005W -1.643913e-01 -1.847008e-03 -0.1483587467 -0.0681185581 -0.0386918053
#> YBR083W -4.074226e-03 -1.008756e-02 -0.0202714179 -0.0061200368  0.0009972112
#> YBR240C -1.312802e-02 -3.144999e-03  0.0403056453 -0.0076283606  0.0128659744
#> YCR106W  2.363483e-02  5.575047e-03 -0.0138394845 -0.0049951220 -0.0263191111
#> YDL056W -1.573310e-02 -1.860046e-02 -0.0903240868  0.0161840356  0.0004353039
#> YDL106C -5.347242e-02  1.203688e-02  0.0868446719  0.0351270915  0.0630143641
#> YDR216W  1.773434e-03 -1.781547e-03  0.0076690369 -0.0125441743 -0.0170057795
#> YDR259C -6.074646e-06  2.001376e-02 -0.0852286405  0.0055841723 -0.0095253382
#> YDR423C -9.650129e-03 -9.825120e-03  0.0738659700 -0.0205307451  0.0145709667
#> YDR463W  4.369652e-02 -5.377891e-02  0.2221736262 -0.0540976027  0.0110012197
#> YEL009C -6.070510e-03 -6.109212e-03 -0.0134549126 -0.0034322501 -0.0043183442
#> YER040W -2.253447e-03 -4.185205e-04 -0.0023955691 -0.0088198783  0.0052776051
#> YFR034C  1.150748e-03  9.643882e-03 -0.0396205547  0.0074575278 -0.0055729196
#> YGL013C -1.443945e-02  4.068630e-03  0.1034703875  0.0061601167  0.0194530869
#> YGL162W  1.229348e-02 -9.385802e-03  0.0114184234 -0.0128136060 -0.0081310783
#> YHL020C  3.761030e-03 -3.683352e-02  0.1332434769 -0.0255400703  0.0441594283
#> YHR084W  7.317209e-04  6.709671e-03  0.0482338410  0.0090137800  0.0033684864
#> YHR178W -1.321982e-02  2.674635e-04 -0.0413912087 -0.0042385472  0.0072334302
#> YHR206W -4.579669e-03 -4.452890e-04 -0.0032016827  0.0026499672 -0.0024228069
#> YIL101C -1.714136e-02 -1.809722e-02 -0.1436630302  0.0127536901  0.0047091876
#> YIL131C  0.000000e+00 -1.073849e-02  0.0265145718  0.0027744357 -0.0030428397
#> YIR018W -6.074646e-06  0.000000e+00 -0.0852286405  0.0055841723 -0.0095253382
#> YJL056C  4.129407e-02 -1.841293e-02  0.0000000000 -0.0304980799  0.0474807655
#> YJL110C  1.117890e-02  4.804594e-03 -0.0278763158  0.0000000000 -0.0123107179
#> YJR060W  2.349301e-02 -1.999667e-02  0.1716871918 -0.0329174832  0.0000000000
#> YKL038W  2.832085e-03 -9.590663e-03 -0.0129850877 -0.0008673974  0.0066146562
#> YKL043W  4.490548e-03 -8.766494e-04 -0.0020477734 -0.0008162026  0.0031585524
#> YKL062W  3.320291e-02  3.366246e-05  0.0912150169 -0.0289814054 -0.0171646138
#> YKL109W -5.545568e-03 -9.571792e-04  0.0422447881 -0.0038073341  0.0041643848
#> YKR099W  3.172355e-03 -6.642811e-03  0.0013588893 -0.0090778643 -0.0065536556
#> YLR131C  1.575858e-02  2.247435e-02  0.0729889876  0.0016120589 -0.0101780702
#> YLR182W  2.061052e-03 -1.133557e-02  0.0162137733  0.0087062492  0.0062312155
#> YLR228C  2.638640e-02 -5.912152e-02  0.2969868991  0.0297553777 -0.0942850113
#> YLR403W -1.740004e-04  5.142883e-03 -0.0095326026  0.0067027375 -0.0081258416
#> YLR451W  1.039765e-01  6.970872e-02  0.0290054876 -0.0369770741 -0.0645571635
#> YML051W -4.081379e-02  1.828614e-02 -0.0197824583  0.0252062749  0.0488269746
#> YMR016C -6.560168e-03  9.755221e-03 -0.0502131260  0.0064265030 -0.0137036368
#> YMR019W -7.508745e-04  2.401556e-02  0.1363438658 -0.0170823308 -0.0084197484
#> YMR043W -2.375084e-03 -3.680956e-03  0.0067046968 -0.0221806866 -0.0118465716
#> YMR070W -6.357745e-03  6.723144e-03 -0.0022966514  0.0013203426 -0.0090117141
#> YNL068C -7.345151e-03  1.157747e-02 -0.0385646293  0.0164834915 -0.0016748116
#> YNL216W  1.200183e-03 -1.722546e-03  0.0001216298  0.0079695062  0.0261734031
#> YOL028C  2.357619e-03  7.994221e-03  0.0679348751  0.0020757559 -0.0018565658
#> YOR113W -2.591806e-02 -8.245089e-03 -0.0136028636  0.0137110953  0.0226996457
#> YOR344C -1.293609e-02 -1.437563e-04 -0.0622587476  0.0187024277  0.0200769859
#> YPL038W  1.892091e-02 -1.485165e-02  0.0318387345 -0.0233488659 -0.0076388752
#> YPL049C -3.878349e-02  1.561530e-03 -0.0676501427 -0.0086420343 -0.0121361545
#> YPL075W -2.096727e-03 -2.522902e-03  0.0649203460  0.0088462927  0.0141923145
#> YPL089C  1.456637e-02  5.755546e-02 -0.1270134864 -0.0004112244 -0.0272066216
#> YPL139C  1.770710e-04  3.389149e-03  0.0503468813 -0.0001526612 -0.0009696360
#> YPL202C -1.561334e-02 -4.670069e-04  0.0092988651  0.0029808368  0.0089558228
#> YPR065W  4.094169e-03 -4.337653e-03  0.0356965170  0.0023867763  0.0124036778
#> YPR104C  7.508807e-03 -1.295211e-02  0.0553166374 -0.0216095826 -0.0178948734
#>               YKL038W       YKL043W       YKL062W       YKL109W       YKR099W
#> YBL005W -3.833053e-02 -0.0667392623  0.0521671100 -0.0346765864  0.0380113373
#> YBR083W  5.116699e-05 -0.0055257811 -0.0206165455 -0.0104308279 -0.0029248353
#> YBR240C  2.119031e-03  0.0017082203  0.0291626383  0.0037353387  0.0108899070
#> YCR106W -7.206220e-03 -0.0146524896 -0.0731422047 -0.0088437952 -0.0275021054
#> YDL056W  7.073424e-03  0.0012854112  0.0026230096 -0.0183634154  0.0056896044
#> YDL106C  4.936260e-02  0.0202649076  0.0811961278 -0.0029673444  0.0234930516
#> YDR216W -1.866310e-03 -0.0003770212 -0.0347702361 -0.0090657251 -0.0122510869
#> YDR259C  1.635600e-03  0.0027249886  0.0100667343 -0.0064072872  0.0055926158
#> YDR423C -1.013578e-03 -0.0031973470 -0.0143769924  0.0094616534 -0.0058543843
#> YDR463W  3.424761e-02 -0.0027948705 -0.2030036948 -0.0280758468 -0.0760962021
#> YEL009C  6.894979e-04  0.0002503473 -0.0009629168 -0.0091172406  0.0014999843
#> YER040W -1.260042e-04 -0.0013700432 -0.0015190878 -0.0031741800  0.0028902727
#> YFR034C -1.923202e-03 -0.0035180553  0.0040423158 -0.0075409144  0.0008112006
#> YGL013C  1.781909e-02  0.0118696911  0.0263962149  0.0111452392  0.0049429298
#> YGL162W -3.678813e-03 -0.0068552562 -0.0306952350 -0.0062339400 -0.0059737965
#> YHL020C  1.616542e-03 -0.0003916243 -0.0061518617  0.0032942353 -0.0004769904
#> YHR084W  4.110570e-03  0.0086254842  0.0047026589  0.0036038984 -0.0057026258
#> YHR178W  6.683316e-03 -0.0028105516  0.0197593343 -0.0136555097  0.0068149011
#> YHR206W  6.976214e-05  0.0026282120  0.0167160426 -0.0026232024  0.0038353463
#> YIL101C  1.348525e-02  0.0081899684  0.0467084168 -0.0110833383  0.0214704098
#> YIL131C -9.638368e-04 -0.0045884722 -0.0013957345 -0.0108181338  0.0024347342
#> YIR018W  1.635600e-03  0.0027249886  0.0100667343 -0.0064072872  0.0055926158
#> YJL056C -5.930491e-02 -0.0652269571  0.0439909393  0.0575999023  0.0002496608
#> YJL110C -2.057469e-03 -0.0021931565 -0.0106524357 -0.0003336874 -0.0085253477
#> YJR060W -2.410555e-04 -0.0004769806 -0.0608894660  0.0007707890 -0.0233860187
#> YKL038W  0.000000e+00  0.0023539309  0.0092700318  0.0018192722  0.0098158163
#> YKL043W -3.392500e-03  0.0000000000  0.0038365231  0.0012119076  0.0050104859
#> YKL062W -7.515415e-04 -0.0099142968  0.0000000000 -0.0039114683 -0.0458705928
#> YKL109W  5.712505e-03  0.0070784755 -0.0036009580  0.0000000000 -0.0039109828
#> YKR099W -5.713720e-04 -0.0023154105 -0.0282752610 -0.0075581080  0.0000000000
#> YLR131C -6.847421e-03 -0.0070271977 -0.0381258254  0.0179960067 -0.0251457906
#> YLR182W -2.389054e-03  0.0035220942  0.0066487699 -0.0012280241  0.0047231664
#> YLR228C -3.063898e-02 -0.0178004508 -0.0193279252 -0.0072621818 -0.0137409977
#> YLR403W -9.575894e-03 -0.0051722543 -0.0054270045  0.0034742192 -0.0058868649
#> YLR451W -3.066638e-02 -0.0448234571 -0.2361920449  0.0261396824 -0.1028976653
#> YML051W  3.825993e-03  0.0037027507  0.0847831293  0.0215857416  0.0300306169
#> YMR016C -7.197652e-03 -0.0034102155 -0.0041028828  0.0035570544  0.0017151846
#> YMR019W  3.949492e-03 -0.0171702062 -0.0202109845 -0.0100636730 -0.0079634782
#> YMR043W -1.323429e-04  0.0067303183 -0.0397151302  0.0071707232 -0.0196885050
#> YMR070W -3.735840e-03 -0.0017811244 -0.0029599572  0.0042728562 -0.0011540869
#> YNL068C  5.562591e-03  0.0059981825  0.0361317667  0.0029389847  0.0137828297
#> YNL216W -1.691368e-03 -0.0054867409  0.0420418207  0.0077943665  0.0154342639
#> YOL028C -2.808069e-03 -0.0013286628 -0.0052462669  0.0019034011 -0.0018296909
#> YOR113W  9.156501e-03  0.0057640251  0.0591283523  0.0047512480  0.0276401712
#> YOR344C  6.851202e-03  0.0062932221  0.0512813742  0.0020038796  0.0175344980
#> YPL038W -9.576268e-03 -0.0071186549 -0.0671531096 -0.0075666013 -0.0176500327
#> YPL049C -8.546279e-03 -0.0134984571 -0.0168594680  0.0006964436 -0.0026954846
#> YPL075W  9.642613e-03  0.0009742530  0.0090162065 -0.0007960092  0.0012043437
#> YPL089C -9.238149e-04 -0.0184135138 -0.0244782414  0.0039189155 -0.0230525977
#> YPL139C  2.679593e-03  0.0012138010 -0.0072160605  0.0041755781 -0.0050712707
#> YPL202C  5.812454e-04  0.0047810399  0.0187306501  0.0074178782  0.0039925072
#> YPR065W  5.501395e-03  0.0039474884  0.0126000937  0.0076240098  0.0061973162
#> YPR104C  7.375946e-03  0.0101786521 -0.0481263634 -0.0160586100 -0.0162395919
#>               YLR131C      YLR182W       YLR228C       YLR403W       YLR451W
#> YBL005W  0.0645157495  0.077553775 -0.3666358147 -0.1734365478 -0.1404260619
#> YBR083W -0.0203573366  0.004145513 -0.0521090705 -0.0016884533 -0.0177613370
#> YBR240C -0.0202273562  0.024128458 -0.0385883812 -0.0073091036 -0.0100422704
#> YCR106W -0.0291473681  0.003561806  0.0370312848  0.0015889400 -0.0208541819
#> YDL056W  0.0132971243 -0.043983482 -0.0250026072  0.0140974081  0.0511143324
#> YDL106C  0.0433005972 -0.014165855 -0.0271853412  0.0111728725 -0.0247463817
#> YDR216W -0.0194532635 -0.032947051 -0.0004942737 -0.0089358394 -0.0087140562
#> YDR259C  0.0104958656 -0.011107537 -0.0164946631  0.0013966531  0.0350157725
#> YDR423C -0.0360659296  0.004058011 -0.0081018261 -0.0011761057 -0.0868375977
#> YDR463W -0.0808136420 -0.149109893 -0.0003109218  0.0373795533 -0.1528933919
#> YEL009C -0.0137543643 -0.014277940 -0.0344379689 -0.0048306152 -0.0030122659
#> YER040W -0.0119147171 -0.002532664 -0.0535377130 -0.0020629730 -0.0169085746
#> YFR034C  0.0039290908  0.017634378 -0.0020922751 -0.0073189467  0.0127276163
#> YGL013C  0.0102087507  0.012643141  0.0235159022  0.0067981308 -0.0123970849
#> YGL162W -0.0285225873 -0.026940485 -0.0272077311 -0.0109106793 -0.0146610968
#> YHL020C -0.0474946479 -0.019314067 -0.0786282156  0.0134259118 -0.0213818961
#> YHR084W  0.0246702914 -0.001687709  0.0643942877  0.0019566371  0.0214760203
#> YHR178W -0.0308688656  0.056176331 -0.1374186089 -0.0083288252 -0.0415350375
#> YHR206W  0.0195380765  0.003546397 -0.0198696909 -0.0075825930  0.0278368048
#> YIL101C  0.0164098935 -0.036914535 -0.0635846075  0.0027576526  0.0665707521
#> YIL131C  0.0039559343 -0.002395257 -0.0600983043 -0.0050104636  0.0156626516
#> YIR018W  0.0104958656 -0.011107537 -0.0164946631  0.0013966531  0.0350157725
#> YJL056C -0.0075037791  0.129426154  0.1758439625 -0.0107209545  0.1434304814
#> YJL110C  0.0100432403  0.016982266  0.0601752385  0.0036281825  0.0130675704
#> YJR060W -0.0407128657 -0.009826751 -0.0370610644 -0.0007065884 -0.0527328880
#> YKL038W -0.0076825116 -0.005498270 -0.0361185683 -0.0003071840 -0.0034452979
#> YKL043W -0.0053621103  0.006846241  0.0056765475 -0.0004658382 -0.0004236055
#> YKL062W -0.0388211789 -0.017444687  0.0508325299  0.0048049094 -0.0896979732
#> YKL109W  0.0156095023 -0.021616406  0.0451798841  0.0053654553 -0.0102660487
#> YKR099W -0.0200886999 -0.014053699 -0.0268170803 -0.0052322578 -0.0187329788
#> YLR131C  0.0000000000  0.025932123  0.1149118675  0.0117318675 -0.0183420836
#> YLR182W  0.0126102843  0.000000000 -0.0303924460  0.0072895037  0.0105476845
#> YLR228C -0.0082483501 -0.017957753  0.0000000000  0.0628232105  0.3210043632
#> YLR403W  0.0198745846  0.005513121  0.0543083035  0.0000000000  0.0155859550
#> YLR451W -0.0092522882  0.041389682  0.3365062810  0.0088407801  0.0000000000
#> YML051W  0.0213709534  0.071468225 -0.0082963816  0.0025746977  0.0168147305
#> YMR016C  0.0145694754 -0.007755655  0.0385914762 -0.0016707511  0.0156169187
#> YMR019W -0.0321396003  0.007163012 -0.0138182978 -0.0365821748 -0.0238228389
#> YMR043W -0.0242994943 -0.029668291  0.0554880399 -0.0051051953 -0.0546855814
#> YMR070W  0.0076535374 -0.002945069  0.0153889148 -0.0024822114 -0.0085346496
#> YNL068C  0.0235494533 -0.011414766  0.0131830835  0.0010199497  0.0589670357
#> YNL216W  0.0066660407  0.052437247 -0.0789743977  0.0026591577  0.0055686208
#> YOL028C  0.0028573874  0.004513678  0.0406279635 -0.0067515427  0.0137774847
#> YOR113W  0.0161490653 -0.011363374 -0.0613224789  0.0061255233  0.0206155675
#> YOR344C  0.0293149175  0.018259933 -0.0435280947  0.0090452407  0.0237222320
#> YPL038W -0.0437420738 -0.009236228 -0.0536674082 -0.0072260516 -0.0660084862
#> YPL049C -0.0008367698  0.002894052  0.0028262174 -0.0145213089 -0.0158036977
#> YPL075W  0.0100808527  0.003244203 -0.0031973415  0.0119192319  0.0164621634
#> YPL089C  0.0014063726  0.126399342  0.0463894018  0.0058661096 -0.0256978468
#> YPL139C  0.0107764980 -0.017481140  0.0418974068 -0.0011905598 -0.0039666054
#> YPL202C  0.0130057074 -0.006503255  0.0110615874  0.0010902902  0.0112244046
#> YPR065W  0.0092493342 -0.005109745  0.0084456178  0.0117426983  0.0044448418
#> YPR104C -0.0353844431 -0.041051093  0.0085179453 -0.0099982108 -0.0177700294
#>               YML051W       YMR016C       YMR019W       YMR043W       YMR070W
#> YBL005W -0.2276301411 -0.0691100034 -0.0278177299 -0.1083231538 -0.0497774535
#> YBR083W -0.0049567856 -0.0105270755  0.0017867413 -0.0117948781 -0.0057976589
#> YBR240C -0.0111599126 -0.0106744172  0.0134326718  0.0056106926 -0.0074432589
#> YCR106W -0.0050352130 -0.0051509917  0.0042816382 -0.0125155444  0.0029136559
#> YDL056W  0.0014816850  0.0052659105  0.0044599493 -0.0100870437  0.0108062263
#> YDL106C  0.0164311477  0.0218872868 -0.0917321505  0.0252175513  0.0143425070
#> YDR216W -0.0009424884 -0.0026549026  0.0041259653  0.0010969228 -0.0059792388
#> YDR259C -0.0054080541  0.0070396329  0.0114777377  0.0017087504  0.0066060704
#> YDR423C -0.0062714151 -0.0223068345 -0.0232575646  0.0109076089 -0.0123725921
#> YDR463W  0.0672225357 -0.0019567793  0.0207969546 -0.0451524598  0.0005083538
#> YEL009C -0.0147541230 -0.0051808048  0.0013461508  0.0028232406 -0.0038567726
#> YER040W -0.0147267279 -0.0053759653  0.0263491228 -0.0093545534 -0.0030581060
#> YFR034C -0.0159988612  0.0053615563 -0.0031962269 -0.0189396089  0.0023683395
#> YGL013C  0.0216233257  0.0064302884 -0.0154338802  0.0032343356  0.0052148955
#> YGL162W -0.0080536529 -0.0049911549  0.0057153670  0.0012561315 -0.0120307342
#> YHL020C  0.0429330503 -0.0190845555 -0.0077298910  0.0412256194 -0.0094550582
#> YHR084W  0.0153115855  0.0157391697 -0.0094036920  0.0065140157  0.0098909068
#> YHR178W -0.0364537483 -0.0190241729 -0.0307692680 -0.0265639267 -0.0200734630
#> YHR206W  0.0032300102  0.0072223758  0.0112387693 -0.0074327579  0.0046479601
#> YIL101C -0.0007506161  0.0048705731  0.0352363204 -0.0016253960 -0.0043678232
#> YIL131C -0.0128438043 -0.0074636423  0.0053466842 -0.0209334964 -0.0027569376
#> YIR018W -0.0054080541  0.0070396329  0.0114777377  0.0017087504  0.0066060704
#> YJL056C -0.1760947364 -0.0054362102  0.1674159828  0.1097373651  0.0267225912
#> YJL110C  0.0167804799  0.0080005710 -0.0264986788 -0.0013358318  0.0061147669
#> YJR060W  0.0434973096 -0.0090769161  0.0170802511 -0.0118010245 -0.0135954311
#> YKL038W  0.0037160506 -0.0050549728  0.0154077200  0.0039830892 -0.0024952398
#> YKL043W  0.0099409060 -0.0015548676 -0.0040845870  0.0056666960 -0.0011225796
#> YKL062W  0.0310470678 -0.0014080890 -0.0201642632 -0.0184312624  0.0003530581
#> YKL109W  0.0230723577  0.0059762780 -0.0091699716  0.0127235268  0.0035253407
#> YKR099W -0.0007613418 -0.0018399965  0.0005609064 -0.0045260499 -0.0016278902
#> YLR131C  0.0160676007  0.0111784971 -0.0386490226 -0.0101335587  0.0090919922
#> YLR182W  0.0133536823  0.0023047232  0.0007791992 -0.0040823113  0.0059006656
#> YLR228C  0.0127288258 -0.0043870881  0.0788239892  0.0054419479  0.0089711543
#> YLR403W -0.0019071330  0.0038282447 -0.0202253898 -0.0008476854  0.0007797682
#> YLR451W  0.0204428295  0.0008984148 -0.0803341534 -0.0552804145 -0.0105626818
#> YML051W  0.0000000000 -0.0082315492 -0.0094715021 -0.0087014792 -0.0020506680
#> YMR016C -0.0112679611  0.0000000000  0.0014764201 -0.0009299947  0.0053507337
#> YMR019W -0.0283148734 -0.0089937537  0.0000000000 -0.0221430415 -0.0347595550
#> YMR043W  0.0073399649 -0.0051701101 -0.0104190681  0.0000000000 -0.0086456389
#> YMR070W  0.0005000042  0.0009693282 -0.0216591874 -0.0018163505  0.0000000000
#> YNL068C -0.0139220314  0.0099319931  0.0124733526  0.0063147943  0.0042259067
#> YNL216W -0.0148880704 -0.0069121628  0.0069720896 -0.0020574128 -0.0087098403
#> YOL028C  0.0015967240  0.0034633294  0.0088093167 -0.0067677585 -0.0003195310
#> YOR113W -0.0262733463 -0.0060727296  0.0218333446  0.0111126832 -0.0010835244
#> YOR344C -0.0023091887  0.0033792166 -0.0093413184  0.0015094893  0.0065571514
#> YPL038W  0.0114706009 -0.0105158866 -0.0053193791 -0.0156635371 -0.0048019328
#> YPL049C -0.0198709951 -0.0160224089 -0.0086114062 -0.0146866887 -0.0149611144
#> YPL075W  0.0036190196  0.0031311636  0.0009528493 -0.0002142159  0.0035991015
#> YPL089C -0.0561027709  0.0023344838 -0.0468037540 -0.0414627693 -0.0038363064
#> YPL139C  0.0091096941  0.0040090577 -0.0021500614  0.0032649280  0.0008035614
#> YPL202C -0.0006816821  0.0016215213 -0.0029601142  0.0178463417  0.0023283839
#> YPR065W  0.0209907144  0.0009820760  0.0120549516  0.0082518904  0.0044461398
#> YPR104C  0.0313086475  0.0024269165  0.0018677502 -0.0013817636  0.0034302296
#>               YNL068C       YNL216W       YOL028C       YOR113W       YOR344C
#> YBL005W -1.367407e-01 -1.021000e-01 -1.751797e-02 -0.1286296090  0.0345549764
#> YBR083W -3.572993e-03  2.299214e-03 -9.011038e-03  0.0026508077  0.0096919865
#> YBR240C -1.099830e-02 -1.670415e-02 -3.490000e-03  0.0131748509  0.0114879950
#> YCR106W  3.401060e-02 -1.059681e-03 -1.385134e-02  0.0244447676 -0.0108929690
#> YDL056W -3.797675e-02  3.270274e-02  1.995176e-03 -0.0328932467 -0.0052887168
#> YDL106C -8.165145e-02  2.516251e-02  1.096560e-02 -0.0613850767  0.0323206196
#> YDR216W  4.100152e-03 -2.981948e-03 -2.919431e-03  0.0104285189 -0.0121350110
#> YDR259C  9.580451e-04  6.652198e-04  8.014583e-03 -0.0085164347 -0.0062379545
#> YDR423C -1.178272e-02  4.005507e-05 -1.434285e-02  0.0093164976  0.0061772567
#> YDR463W  1.403435e-02  1.265716e-01 -9.519211e-03  0.0769931823 -0.0002077111
#> YEL009C -9.541243e-03 -4.648981e-03 -7.050548e-04 -0.0066634562 -0.0037892929
#> YER040W  4.239790e-03 -4.020124e-03  1.654211e-04  0.0110195619  0.0071691044
#> YFR034C  1.346719e-03 -1.144260e-02  6.200071e-06 -0.0002229177 -0.0019556266
#> YGL013C -1.730202e-02  1.200068e-03  8.165133e-03 -0.0127771981  0.0066698588
#> YGL162W  6.799887e-03 -3.511581e-03 -7.300833e-04  0.0100164393 -0.0041885054
#> YHL020C -2.795813e-02  4.414670e-02 -1.412809e-02  0.0390909928  0.0181409241
#> YHR084W -5.105041e-04  2.143866e-03  3.449183e-03 -0.0031506879 -0.0005939308
#> YHR178W -1.481361e-02 -2.055498e-02 -2.359737e-02 -0.0018553418  0.0311830647
#> YHR206W -2.843113e-03 -3.230348e-03  6.562803e-03 -0.0033265563 -0.0032505072
#> YIL101C -3.501674e-02  1.801454e-02  1.972783e-02 -0.0329927133 -0.0061263517
#> YIL131C -1.181310e-02 -2.842717e-03 -4.570946e-03 -0.0099781671  0.0050456519
#> YIR018W  9.580451e-04  6.652198e-04  8.014583e-03 -0.0085164347 -0.0062379545
#> YJL056C  3.348766e-02 -4.656374e-02  4.168788e-02  0.0231045611 -0.0375762774
#> YJL110C  1.031894e-02 -2.047059e-04 -3.021456e-03 -0.0006691651 -0.0084637638
#> YJR060W  2.750647e-02  1.705263e-02 -8.428741e-03  0.0542496771  0.0106557359
#> YKL038W -1.706299e-03  3.879469e-03 -1.283630e-03  0.0054712319  0.0057703402
#> YKL043W  5.346836e-03 -3.055245e-03 -2.133544e-03  0.0062016460  0.0029912353
#> YKL062W  4.231719e-02  1.974385e-02 -1.420352e-02  0.0442703409 -0.0066103261
#> YKL109W -3.329478e-03  1.060791e-02  4.049258e-03 -0.0029689872 -0.0022736640
#> YKR099W -2.894512e-05  3.448056e-03 -3.489027e-03  0.0119551840 -0.0052164481
#> YLR131C  2.885573e-02 -9.640220e-04 -8.984471e-03  0.0116688978 -0.0001546435
#> YLR182W -7.681034e-03  1.487630e-02  8.228543e-04 -0.0069840895  0.0026700224
#> YLR228C  4.756716e-02  2.315150e-03  7.026362e-03  0.0102157262 -0.0700188910
#> YLR403W  2.782799e-03 -4.848479e-03 -4.037220e-03 -0.0018357729 -0.0044987095
#> YLR451W  1.643094e-01 -9.722886e-03 -2.428164e-02  0.0569273892 -0.0277870627
#> YML051W -3.634635e-02 -2.770180e-02 -2.686436e-03 -0.0537645854  0.0316117920
#> YMR016C  3.728600e-03 -7.919340e-03  3.454990e-03 -0.0110232336 -0.0079389236
#> YMR019W  1.038737e-02 -5.144802e-02 -4.395708e-03  0.0175121310 -0.0030517067
#> YMR043W  2.528373e-03  6.313572e-03 -7.584140e-03  0.0166550476 -0.0080129964
#> YMR070W -4.422013e-03 -6.496937e-03 -1.249067e-03 -0.0087172704 -0.0023176606
#> YNL068C  0.000000e+00 -3.183586e-03  1.433817e-02 -0.0207935800 -0.0094995185
#> YNL216W -3.426852e-03  0.000000e+00  7.842733e-04 -0.0046382581  0.0228157975
#> YOL028C  1.068470e-02 -9.522618e-03  0.000000e+00  0.0110710432  0.0002409193
#> YOR113W -3.812182e-02  3.125221e-03  1.413270e-02  0.0000000000  0.0067594619
#> YOR344C -2.527768e-02  9.021033e-03  4.225885e-03 -0.0308112043  0.0000000000
#> YPL038W  2.106660e-02  4.598689e-03 -1.665184e-02  0.0355786933  0.0042689398
#> YPL049C -1.550656e-02 -2.310980e-02 -3.627436e-03 -0.0215514070  0.0018201966
#> YPL075W -5.289720e-03  2.015023e-02  9.117936e-03 -0.0080152926  0.0063442964
#> YPL089C  6.414114e-02 -4.903802e-02 -1.352133e-02  0.0109805546  0.0141042832
#> YPL139C  8.054637e-05  3.934613e-03  5.603578e-03  0.0003639808 -0.0062632380
#> YPL202C -1.774306e-02  3.055596e-04  3.821703e-03 -0.0102337970  0.0016552669
#> YPR065W  1.248023e-03  1.810674e-02  8.398313e-03 -0.0057813902  0.0014313393
#> YPR104C  8.321161e-03  6.992380e-03 -1.310410e-02  0.0239478018 -0.0131055484
#>               YPL038W       YPL049C       YPL075W       YPL089C       YPL139C
#> YBL005W  0.0332402099 -0.1226100894 -0.0892895396  0.0343337001 -0.1073232608
#> YBR083W -0.0029006987 -0.0109465812  0.0039655441 -0.0085656350 -0.0033731960
#> YBR240C  0.0059487075 -0.0184344711 -0.0016763205 -0.0458644946  0.0026263737
#> YCR106W -0.0287813559  0.0172782278 -0.0107097024 -0.0003520843  0.0087481560
#> YDL056W  0.0113081712  0.0196371640 -0.0058325251  0.0588097391 -0.0081462303
#> YDL106C  0.0911756278 -0.0037287426  0.0047597283 -0.0155834444 -0.0161339939
#> YDR216W -0.0176950964  0.0011647130 -0.0053684993  0.0018745902 -0.0031303387
#> YDR259C -0.0015011940  0.0032379155 -0.0093059990  0.0211214517  0.0020547353
#> YDR423C  0.0107430419 -0.0189062855  0.0060929308 -0.0562745998 -0.0065739557
#> YDR463W -0.0182528833  0.0336685328 -0.0315868124 -0.0495249080  0.0021661423
#> YEL009C -0.0015290880 -0.0017832278 -0.0131371249 -0.0028461771 -0.0039593698
#> YER040W -0.0023123481 -0.0129586916 -0.0045509389 -0.0354369630  0.0023174270
#> YFR034C  0.0020695132  0.0021254793 -0.0033270201  0.0127303532 -0.0030171727
#> YGL013C  0.0245997973 -0.0040348654  0.0092699550 -0.0301908837  0.0045979676
#> YGL162W -0.0231329073  0.0114843110 -0.0060240973  0.0219120833 -0.0004253251
#> YHL020C  0.0266034707  0.0197906910  0.0323189170  0.0137590862 -0.0051917760
#> YHR084W  0.0012853725  0.0133282218 -0.0013435787  0.0127631940  0.0053792022
#> YHR178W  0.0084850511 -0.0407994650 -0.0051283325 -0.0703554737 -0.0111753748
#> YHR206W  0.0083969977 -0.0006940119  0.0012157338  0.0159222547 -0.0054321674
#> YIL101C  0.0192525226  0.0039072432 -0.0179986735  0.0327467175 -0.0071437719
#> YIL131C -0.0028403344 -0.0160359916 -0.0011333662  0.0010678106 -0.0084087849
#> YIR018W -0.0015011940  0.0032379155 -0.0093059990  0.0211214517  0.0020547353
#> YJL056C -0.0412736335  0.0243727919 -0.0257787200  0.0318312752  0.0604930054
#> YJL110C -0.0029219233  0.0102040262  0.0018625200  0.0239534027 -0.0004227661
#> YJR060W -0.0198922646  0.0005795224  0.0094251829 -0.0412917586  0.0071138787
#> YKL038W -0.0029120297  0.0008057411  0.0062898975  0.0057951759  0.0017006440
#> YKL043W  0.0002395495  0.0026557630  0.0026704419  0.0005670048  0.0009886025
#> YKL062W -0.0292614653  0.0054121307 -0.0116458932 -0.0403114215 -0.0002553759
#> YKL109W  0.0054373029  0.0008219535 -0.0001732692 -0.0073153643  0.0029417798
#> YKR099W -0.0037301224  0.0031875871 -0.0081771059  0.0028786801 -0.0044256266
#> YLR131C -0.0192098745  0.0065838297  0.0028937552 -0.0162618700  0.0085467317
#> YLR182W  0.0033005121  0.0077995347  0.0059812264  0.0293601338 -0.0048179767
#> YLR228C -0.1587707570  0.1048731138 -0.0067190382  0.1063333823  0.0410840604
#> YLR403W -0.0062410702  0.0039003047  0.0016572875  0.0260935684 -0.0049844593
#> YLR451W -0.0830516496 -0.0114437533 -0.0079583114 -0.0889828331  0.0014127692
#> YML051W  0.0277409186 -0.0179319520  0.0079915610 -0.0557168336  0.0027487291
#> YMR016C -0.0088794644 -0.0037442714 -0.0082350636  0.0157478630  0.0014852734
#> YMR019W -0.0163822831 -0.0113901473  0.0045779327 -0.0568419656 -0.0083921410
#> YMR043W -0.0115432584 -0.0036063153 -0.0062335231 -0.0317763108  0.0051842918
#> YMR070W -0.0027658678 -0.0033170490 -0.0062247676  0.0028950908 -0.0016630156
#> YNL068C  0.0056523743  0.0083821512  0.0043301353  0.0304652444  0.0027946633
#> YNL216W  0.0104652777 -0.0172385391  0.0280055370 -0.0022621899 -0.0005781312
#> YOL028C -0.0101644007  0.0024941988  0.0092656552  0.0020339168  0.0049640284
#> YOR113W  0.0249196512 -0.0060474378  0.0009466043  0.0009222938  0.0007915064
#> YOR344C  0.0270546966 -0.0045959942  0.0127779621  0.0186032949 -0.0091420422
#> YPL038W  0.0000000000  0.0068470900 -0.0006452053 -0.0068472857  0.0009032306
#> YPL049C -0.0002459688  0.0000000000 -0.0068138151 -0.0290663120 -0.0124736504
#> YPL075W  0.0147003393  0.0043643805  0.0000000000  0.0113195285  0.0024115727
#> YPL089C -0.0074073936 -0.0588775067 -0.0012031100  0.0000000000  0.0145530578
#> YPL139C  0.0027813623  0.0027041555  0.0018546029  0.0036100012  0.0000000000
#> YPL202C  0.0094430232 -0.0020013849  0.0030024128  0.0079034927  0.0004277757
#> YPR065W  0.0081990313  0.0075672216  0.0140933561  0.0040837373  0.0033041454
#> YPR104C -0.0192330159  0.0218494415 -0.0198073762 -0.0236812113  0.0023242803
#>               YPL202C       YPR065W       YPR104C
#> YBL005W -0.1078045812 -3.513803e-02 -0.0605950775
#> YBR083W -0.0039656396  6.095964e-04  0.0001619492
#> YBR240C -0.0061327032  6.449845e-03 -0.0095231552
#> YCR106W -0.0021798003 -1.342038e-02 -0.0112030532
#> YDL056W -0.0061839473 -1.475223e-02  0.0172668014
#> YDL106C  0.0135835828  4.592580e-03  0.0154503139
#> YDR216W -0.0027441088 -3.002084e-03 -0.0045065715
#> YDR259C -0.0021578012 -4.585649e-03 -0.0024222663
#> YDR423C  0.0001421087  3.147118e-03  0.0167525784
#> YDR463W  0.0282238902 -3.670612e-02  0.0768663086
#> YEL009C -0.0076084857 -3.514642e-03 -0.0010119932
#> YER040W -0.0020364791 -2.863899e-03 -0.0054277028
#> YFR034C -0.0044305045 -5.250822e-03 -0.0165298810
#> YGL013C  0.0053428087  1.153551e-02  0.0070639099
#> YGL162W  0.0033228399 -7.908824e-03 -0.0083125311
#> YHL020C  0.0198799617 -2.987537e-05  0.0293187679
#> YHR084W  0.0065254340  3.598211e-03 -0.0069305984
#> YHR178W -0.0084835261  7.027596e-03 -0.0188895829
#> YHR206W -0.0083963615  4.237976e-03 -0.0035891838
#> YIL101C -0.0065420538 -1.400748e-02  0.0151194910
#> YIL131C -0.0140710219 -1.440490e-05 -0.0050682477
#> YIR018W -0.0021578012 -4.585649e-03 -0.0024222663
#> YJL056C  0.0022292105 -1.291391e-02  0.0045374326
#> YJL110C  0.0009111622 -4.559457e-04 -0.0006042229
#> YJR060W  0.0165806307  4.265932e-03  0.0043154177
#> YKL038W  0.0003257613  1.855928e-03  0.0073516949
#> YKL043W  0.0034141735  4.364432e-05  0.0002729673
#> YKL062W  0.0107716948 -3.932989e-03  0.0138097158
#> YKL109W  0.0088708280  1.840371e-03  0.0101537659
#> YKR099W -0.0028190165 -4.691945e-03  0.0004977904
#> YLR131C  0.0123769819  4.317992e-03 -0.0050829772
#> YLR182W -0.0009115289  2.597365e-03  0.0150963905
#> YLR228C -0.0159151578  2.314708e-02 -0.0085816897
#> YLR403W  0.0009229392 -1.081808e-04 -0.0104645634
#> YLR451W  0.0263220266 -7.251113e-03  0.0103890163
#> YML051W -0.0127582894  2.365432e-02 -0.0049170539
#> YMR016C -0.0033215680 -6.053277e-03 -0.0046615924
#> YMR019W -0.0160783682  1.036725e-02 -0.0414189532
#> YMR043W  0.0156020454 -1.031020e-03  0.0040540440
#> YMR070W -0.0016435433  1.791018e-03  0.0008921531
#> YNL068C -0.0052680545  2.687627e-03 -0.0035528000
#> YNL216W  0.0021891924  9.227155e-03 -0.0061257215
#> YOL028C  0.0010521105  4.250116e-03 -0.0169054496
#> YOR113W -0.0072950966  4.329804e-04  0.0135482905
#> YOR344C -0.0056336984  6.482611e-03  0.0155229853
#> YPL038W  0.0086503733 -4.528936e-03 -0.0018152563
#> YPL049C -0.0156968094 -7.540144e-04 -0.0048034710
#> YPL075W  0.0033678326  8.090804e-03  0.0107456702
#> YPL089C  0.0042925856  9.362122e-03 -0.0535871733
#> YPL139C  0.0015125360  3.317066e-04  0.0054605545
#> YPL202C  0.0000000000 -1.114375e-03 -0.0018256638
#> YPR065W  0.0028970833  0.000000e+00  0.0251863711
#> YPR104C -0.0074259460  4.416047e-03  0.0000000000