Run python implementation of LIONESS

LIONESS(Linear Interpolation to Obtain Network Estimates for Single Samples) is a method to estimate sample-specific regulatory networks. [(LIONESS arxiv paper)]).

lioness.py(expr = expression, motif = motif, ppi = ppi,
  mode_process = "union", rm_missing = FALSE, start_sample = 1,
  end_sample = "None", save_dir = "lioness_output", save_fmt = "npy")

Arguments

expr
Character string indicating the file path of expression values file, as each gene (row) by samples (columns) required

motif
Character string indicating the file path of pair file of motif edges, when not provided, analysis continues with Pearson correlation matrix. optional

ppi
Character string indicating the pair file path of Protein-Protein interaction dataset. optional

mode_process
An optional character to define the pre-processing of input dataset, options include "legacy" to the original behavior of PANDA in Python implement, "interaction" to constrain the TF and gene as interaction across all input datasets, and "union" to take TF union and gene union across all input datasets. The default value is "union".

rm_missing
When mode_process = "legacy", rm_missing is an optional boolean indicatining whether to remove genes and tfs not present in all input files. If TRUE, remove all unmatched tf and genes. if FALSE, keep all tf and genes. The default value is FALSE.

start_sample
Numeric indicating the start sample number, The default value is 1.

end_sample
Numeric indicating the start sample number, The default value is None. optional

save_dir
Character string indicating the folder name of output lioness networks for each sample by defined.optional The default is a folder named "lioness_output" under current working directory.

save_fmt
Character string indicating the format of lioness network of each sample. The dafault is "npy". The option is txt, npy, or mat.

Value

A data frame with columns representing each sample, rows representing the regulator-target pair in PANDA network generated by runPanda. Each cell filled with the related score, representing the estimated contribution of a sample to the aggregate network.

Examples

# refer to the input datasets files of control in inst/extdat as example
control_expression_file_path <- system.file("extdata", "expr10_matched.txt", package = "netZooR", mustWork = TRUE)
motif_file_path <- system.file("extdata", "chip_matched.txt", package = "netZooR", mustWork = TRUE)
ppi_file_path <- system.file("extdata", "ppi_matched.txt", package = "netZooR", mustWork = TRUE)

# Run LIONESS algorithm
control_lioness_result <- lioness.py(expr = control_expression_file_path, motif = motif_file_path, ppi = ppi_file_path, mode_process="legacy", rm_missing = TRUE,start_sample=1, end_sample=2)

#> Use the legacy mode to pre-processing dataset and keep the matched tfs or genes

#> Rename the context of first two columns with prefix 'reg_' and 'tar_'