MOdeling Network State Transitions from Expression and Regulatory data (MONSTER)

This function runs the MONSTER algorithm. Biological states are characterized by distinct patterns of gene expression that reflect each phenotype's active cellular processes. Driving these phenotypes are gene regulatory networks in which transcriptions factors control when and to what degree individual genes are expressed. Phenotypic transitions, such as those that occur when disease arises from healthy tissue, are associated with changes in these networks. MONSTER is an approach to understanding these transitions. MONSTER models phenotypic-specific regulatory networks and then estimates a "transition matrix" that converts one state to another. By examining the properties of the transition matrix, we can gain insight into regulatory changes associated with phenotypic state transition.

monster(expr, design, motif, nullPerms = 100, ni_method = "bere",
  ni.coefficient.cutoff = NA, numMaxCores = 1, outputDir = NA)

Arguments

expr
Gene Expression dataset, can be matrix or data.frame of expression values or ExpressionSet

design
Binary vector indicating case control partition

motif
Regulatory data.frame consisting of three columns. For each row, a transcription factor (column 1) regulates a gene (column 2) with a defined strength (column 3), usually taken to be 0 or 1

nullPerms
number of random permutations to run (default 100). Set to 0 to only calculate observed transition matrix

ni_method
String to indicate algorithm method. Must be one of "bere","pearson","cd","lda", or "wcd". Default is "bere"

ni.coefficient.cutoff
numeric to specify a p-value cutoff at the network inference step. Default is NA, indicating inclusion of all coefficients.

numMaxCores
requires doParallel, foreach. Runs MONSTER in parallel computing environment. Set to 1 to avoid parallelization.

outputDir
character vector specifying a directory or path in which which to save MONSTER results, default is NA and results are not saved.

Value

An object of class "monsterAnalysis" containing results

Examples

# \donttest{
data(yeast)
design <- c(rep(0,20),rep(NA,10),rep(1,20))
monsterRes <- monster(yeast$exp.cc[1:500,], design, yeast$motif, nullPerms=10, numMaxCores=4)

#> [1] "Running null permutations in parallel"
#> [1] "4 cores used"
#> [1] "11 network transitions to be estimated"
#> Time difference of 10.7082 secs

plot(monsterRes)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

# }