Run EGRET in R

Description: NOTE: Beta version. EGRET infers individual-specific gene regulatory networks using inidividual level data - a genotype vcf file (v) and QBiC binding predictions (q) - as well as population/reference level data - eQTLs (b), a motif-gene prior (m), PPI network (p), and gene expression (e). An annotation file g is also used to map TF names to their corresponding ensemble ids.

runEgret(b, v, q, m, e, p, g, t)

Arguments

b
: Data frame of eQTL data, each row containing an eQTL which exist within motif regions adjacent to the eGene, with columns TF, gene, variant position,variant chromosome, eQTL beta value.

v
: Data frame of VCF file containing SNPs of the individual in question

q
: Data frame of QBiC predictions of the effect of eQTL variants on TF binding. Each row represents an eQTL variant with a predicted negative (disruptive) effect on the binding of the TF corresponding to the motif in which the eQTL variant resides. Colums are: eQTL variant as chr[chrNum]_position, TF, adjacent eGene, QBiC binding effect size and QBiC binding effect (should be negative)

m
: Motif prior data frame. Each row represents an edge in the bipartite motif prior, with columns TF, gene and edge weight. The edge weight should be 1 or 0 based on the presence/absence of the TF motif in the promoter region of the gene.

e
: Gene expression data frame in which each row represents a gene and each column represents the expression of that gene in a sample. The first column should contain gene IDs.

p
: PPI network data frame. Each row represents an edgem with columns TF, TF and interaction weight.

g
: Data frame mapping gene names to gene ids, with columns containing the gene ID the corresponding gene name.

t
: A string containing a name for the EGRET run. Output files will be labelled with this tag.

Outputs:

Value

EGRET : Predicted genotye-specific gene regulatory network saved as tag_egret.RData

BASELINE : A Baseline (PANDA) genotype-agnostic gene regulatory network saved as tag_panda.RData

Details

Inputs:

Examples

# Run EGRET algorithm
toy_qbic_path <- system.file("extdata", "toy_qbic.txt", package = "netZooR", 
mustWork = TRUE)
toy_genotype_path <- system.file("extdata", "toy_genotype.vcf", 
package = "netZooR", mustWork = TRUE)
toy_motif_path <- system.file("extdata", "toy_motif_prior.txt", 
package = "netZooR", mustWork = TRUE)
toy_expr_path <- system.file("extdata", "toy_expr.txt", 
package = "netZooR", mustWork = TRUE)
toy_ppi_path <- system.file("extdata", "toy_ppi_prior.txt", 
package = "netZooR", mustWork = TRUE)
toy_eqtl_path <- system.file("extdata", "toy_eQTL.txt", 
package = "netZooR", mustWork = TRUE)
toy_map_path <- system.file("extdata", "toy_map.txt", 
package = "netZooR", mustWork = TRUE)
qbic <- read.table(file = toy_qbic_path, header = FALSE)
vcf <- read.table(toy_genotype_path, header = FALSE, sep = "\t", 
stringsAsFactors = FALSE, 
colClasses = c("character", "numeric", "character", "character", "character", 
"character", "character", "character", "character", "character"))
motif <- read.table(toy_motif_path, sep = "\t", header = FALSE)
expr <- read.table(toy_expr_path, header = FALSE, sep = "\t", row.names = 1)
ppi <- read.table(toy_ppi_path, header = FALSE, sep = "\t")
qtl <- read.table(toy_eqtl_path, header = FALSE)
nameGeneMap <- read.table(toy_map_path, header = FALSE)
tag <- "my_toy_egret_run"
# \donttest{
runEgret(qtl,vcf,qbic,motif,expr,ppi,nameGeneMap,tag)

#> [1] "Initializing and validating"

#> 0 genes removed that were not present in motif

#> 12 motif edges removed that targeted genes missing in expression data

#> [1] "Verified sufficient samples"

#> 0 PPI edges removed that were not present in motif

#> [1] "Normalizing networks..."
#> [1] "Learning Network..."
#> [1] "Using tanimoto similarity"

#> Iteration0: hamming distance =0.35159

#> Iteration1: hamming distance =0.28205

#> Iteration2: hamming distance =0.25188

#> Iteration3: hamming distance =0.24416

#> Iteration4: hamming distance =0.25299

#> Iteration5: hamming distance =0.26142

#> Iteration6: hamming distance =0.26807

#> Iteration7: hamming distance =0.26952

#> Iteration8: hamming distance =0.26802

#> Iteration9: hamming distance =0.25899

#> Iteration10: hamming distance =0.24379

#> Iteration11: hamming distance =0.22374

#> Iteration12: hamming distance =0.20081

#> Iteration13: hamming distance =0.17714

#> Iteration14: hamming distance =0.15418

#> Iteration15: hamming distance =0.13283

#> Iteration16: hamming distance =0.11353

#> Iteration17: hamming distance =0.09643

#> Iteration18: hamming distance =0.08151

#> Iteration19: hamming distance =0.06862

#> Iteration20: hamming distance =0.05758

#> Iteration21: hamming distance =0.04819

#> Iteration22: hamming distance =0.04023

#> Iteration23: hamming distance =0.03352

#> Iteration24: hamming distance =0.02789

#> Iteration25: hamming distance =0.02316

#> Iteration26: hamming distance =0.01921

#> Iteration27: hamming distance =0.01591

#> Iteration28: hamming distance =0.01317

#> Iteration29: hamming distance =0.01089

#> Iteration30: hamming distance =0.00899

#> Iteration31: hamming distance =0.00742

#> Iteration32: hamming distance =0.00612

#> Iteration33: hamming distance =0.00504

#> Iteration34: hamming distance =0.00415

#> Iteration35: hamming distance =0.00342

#> Iteration36: hamming distance =0.00281

#> Iteration37: hamming distance =0.00231

#> Iteration38: hamming distance =0.0019

#> Iteration39: hamming distance =0.00156

#> Iteration40: hamming distance =0.00128

#> Iteration41: hamming distance =0.00105

#> Iteration42: hamming distance =0.00087

#> Successfully ran PANDA on 4 Genes and 6 TFs.
#> Time elapsed:0.02seconds.

#> [1] "Initializing and validating"

#> 0 genes removed that were not present in motif

#> 12 motif edges removed that targeted genes missing in expression data

#> [1] "Verified sufficient samples"

#> 0 PPI edges removed that were not present in motif

#> [1] "Normalizing networks..."
#> [1] "Learning Network..."
#> [1] "Using tanimoto similarity"

#> Iteration0: hamming distance =0.22762

#> Iteration1: hamming distance =0.2101

#> Iteration2: hamming distance =0.21795

#> Iteration3: hamming distance =0.25112

#> Iteration4: hamming distance =0.28488

#> Iteration5: hamming distance =0.31193

#> Iteration6: hamming distance =0.33076

#> Iteration7: hamming distance =0.34225

#> Iteration8: hamming distance =0.34137

#> Iteration9: hamming distance =0.32951

#> Iteration10: hamming distance =0.30886

#> Iteration11: hamming distance =0.28222

#> Iteration12: hamming distance =0.25247

#> Iteration13: hamming distance =0.22205

#> Iteration14: hamming distance =0.19274

#> Iteration15: hamming distance =0.1656

#> Iteration16: hamming distance =0.14117

#> Iteration17: hamming distance =0.11962

#> Iteration18: hamming distance =0.10088

#> Iteration19: hamming distance =0.08476

#> Iteration20: hamming distance =0.071

#> Iteration21: hamming distance =0.05932

#> Iteration22: hamming distance =0.04946

#> Iteration23: hamming distance =0.04116

#> Iteration24: hamming distance =0.0342

#> Iteration25: hamming distance =0.02838

#> Iteration26: hamming distance =0.02352

#> Iteration27: hamming distance =0.01947

#> Iteration28: hamming distance =0.0161

#> Iteration29: hamming distance =0.0133

#> Iteration30: hamming distance =0.01098

#> Iteration31: hamming distance =0.00906

#> Iteration32: hamming distance =0.00747

#> Iteration33: hamming distance =0.00615

#> Iteration34: hamming distance =0.00507

#> Iteration35: hamming distance =0.00417

#> Iteration36: hamming distance =0.00343

#> Iteration37: hamming distance =0.00282

#> Iteration38: hamming distance =0.00232

#> Iteration39: hamming distance =0.0019

#> Iteration40: hamming distance =0.00156

#> Iteration41: hamming distance =0.00128

#> Iteration42: hamming distance =0.00105

#> Iteration43: hamming distance =0.00087

#> Successfully ran PANDA on 4 Genes and 6 TFs.
#> Time elapsed:0.02seconds.

# }